api/pyproject.toml
[tool.poetry]
name = "iheroes_api"
version = "0.0.1"
description = "iHeroes Threat Manager API"
authors = ["Giovanni Armane <giovanniarmane@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.2"
argon2-cffi = "^20.1.0"
databases = {version = "^0.3.2", extras = ["postgresql"]}
email-validator = "^1.1.1"
fastapi = {version = "^0.55.1", extras = ["email_validator", "requests"]}
passlib = "^1.7.2"
pyjwt = "^1.7.1"
python-multipart = "^0.0.5"
toolz = "^0.10.0"
uvicorn = "^0.11.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.2"
flake8-black = "^0.2.0"
flake8-broken-line = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.2.3"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^0.4.0"
flake8-pytest-style = "^1.1.1"
mccabe = "^0.6.1"
mypy = "^0.780"
pep8-naming = "^0.10.0"
pytest = "^5.4.3"
pytest-asyncio = "^0.12.0"
pytest-cov = "^2.9.0"
pytest-mock = "^3.1.1"
pytest-sugar = "^0.9.3"
python-dotenv = "^0.13.0"
Faker = "^4.1.0"
pytest-factoryboy = "^2.0.3"
factory_boy = "^2.12.0"

[tool.poetry.scripts]
seeder = "scripts:seeder"
web_server = "scripts:web_server"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
